at boot: stop deleting cannons with invalid owners
If the owner "has not played before", the cannon was being silently
deleted (they could be silently deleted for many other reasons too).
In particular, this caused a problem because I've purged the creating
user from the world/playerdata, but I still need the cannons they
created to function properly.

Silent deletion of data in the database without a user initiating
it seems like it's never the right answer, so I've changed those
log messages from level debug to level severe, but the better
solution is to refactor this to not delete them unless a user
explicitly desires this behavior.